feat: add useOffset hook (#20)
