+++ packages/client/src/routes/index.tsx
@@ -1,5 +1,4 @@
 import { Route, Routes } from 'react-router-dom';
-import CRDTPage from '~/pages/CRDTPage';
 import HomePage from '~/pages/HomePage';
 import NotePage from '~/pages/NotePage';
 import NotFoundPage from '~/pages/NotFoundPage';
@@ -9,7 +8,6 @@ const PageRoutes = () => {
     <Routes>
       <Route index element={<HomePage />} />
       <Route path="/note/:noteId" element={<NotePage />} />
-      <Route path="/crdt" element={<CRDTPage />} />
       <Route path="*" element={<NotFoundPage />} />
     </Routes>
   );
