+++ packages/common/src/crdt/crdt.ts
@@ -25,6 +25,7 @@ export class CRDT {
       .reduce((prev, cur) => Math.max(prev, cur), 0);
 
     this.clock = maxClock + 1;
+    console.log('clock', this.clock);
   }
 
   get data() {
